requirements.txt
Flask-API==0.6.5
Flask-Testing==0.4.2
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
pbr==1.8.1
requests==2.9.1
six==1.10.0
stevedore==1.12.0
virtualenv==14.0.6
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
Werkzeug==0.11.4
wheel==0.29.0
